Kommentointi valmis, ääkköset aiheuttivat vielä hieman päänvaivaa, mutta nyt pitäisi toimia.

+++ Tapahtumatietokanta/src/main/java/servlets/DeleteEventServlet.java
@@ -29,7 +29,7 @@ public class DeleteEventServlet extends MainServlet {
         if(confirmLogin(request, response)) {
             
             try {
-                long eventKey = Integer.parseInt(request.getParameter("event"));
+                long eventKey = Long.parseLong(request.getParameter("event"));
 
                 (new EventQuery()).deleteEvent(eventKey);
